doc:更新数据库脚本文件

+++ youlai-common/common-core/src/main/java/com/youlai/common/core/result/Result.java
@@ -20,7 +20,7 @@ public class Result<T> implements Serializable {
     private String msg;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    private long total;
+    private Integer total;
 
     public static <T> Result<T> success() {
         return success(null);
@@ -67,12 +67,12 @@ public class Result<T> implements Serializable {
         return result;
     }
 
-    public static <T> Result<T> success(T data, long total) {
+    public static <T> Result<T> success(T data, Long total) {
         Result<T> result = new Result();
         result.setCode(ResultCode.SUCCESS.getCode());
         result.setMsg(ResultCode.SUCCESS.getMsg());
         result.setData(data);
-        result.setTotal(total);
+        result.setTotal(total.intValue());
         return result;
     }
